requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.1.0
cattrs==23.1.2
Django==4.2.5
django-extensions==3.2.3
django-ninja==1.1.0
django-schema-graph==3.1.0
djangorestframework==3.14.0
drf-spectacular==0.28.0
exceptiongroup==1.1.3
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pydantic==2.6.3
pydantic_core==2.16.3
pytz==2023.3.post1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
